Move ScoreCardFinder interface to producers package
