+++ scorecards-service/src/main/java/mikaa/feature/results/ScoreEntry.java
@@ -1,4 +1,4 @@
 package mikaa.feature.results;
 
-public record ScoreEntry(long id, long playerId, int hole, int par, int score) {
+record ScoreEntry(long id, long playerId, int hole, int par, int score) {
 }
