+++ scorecards-service/src/main/java/mikaa/feature/scorecard/ScoreCardService.java
@@ -11,6 +11,7 @@ import mikaa.domain.NewScoreCard;
 import mikaa.domain.ScoreCard;
 import mikaa.feature.course.CourseFinder;
 import mikaa.feature.player.PlayerFinder;
+import mikaa.producers.ScoreCardFinder;
 import mikaa.producers.ScoreCardProducer;
 
 @ApplicationScoped
